requirements.txt
google-generativeai
fastapi
uvicorn
requests
AIzaSyDopnKc4gSmepCI6RHtZdsUDh0GqKn8lKs